% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Algorithm To Detect Outliers}
\usage{
detectOutliers(y, X, intercepto = 1)
}
\arguments{
\item{y}{A vector with response variables.}

\item{X}{A matrix or vector with explanatory variables.}

\item{intercepto}{1 for a model with intercept and 0 for a model without intercept.}
}
\value{
\item{outliers     }{          A vector with outliers indices.}.
}
\description{
Algorithm To Detect Outliers
}
\examples{
### Using stackloss data

detectOutliers(stack.loss, stack.x)
}
\references{
Dodge, Y. (1997) Lad regression for detecting outliers in response and explanatory variables.
\emph{Journal of Multivariate Analysis}, \strong{61}, 144–158. \doi{10.1006/jmva.1997.1666}
}
