% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD1.R
\name{FD1}
\alias{FD1}
\title{F1 Distance}
\usage{
FD1(y, X)
}
\arguments{
\item{y}{A vector with response variables.}

\item{X}{A matrix or vector with explanatory variables.}
}
\value{
\item{F1 Distance}{          A vector with F1 Distance for each observation.}.
}
\description{
F1 Distance
}
\examples{
### Using stackloss data

CookDistance(stack.loss, stack.x)
}
\references{
Sun, R.-B. and Wei, B.-C. (2004) On influence assessment for lad regression.
\emph{Statistics & Probability Letters}, \strong{67}, 97-110. \doi{10.1016/j.spl.2003.08.018}.
}
