% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houses.R
\docType{data}
\name{houses}
\alias{houses}
\title{Prices of 27 houses}
\format{A data frame with 27 rows and 5 variables}
\usage{
data(houses)
}
\description{
A dataset containing the prices and other attributes of 27 houses. The variables are as follows:
}
\details{
\itemize{
  \item Price. sale price of house in thousands of US dollars
  \item Taxes. taxes in hundreds of US dollars
  \item Area. total area in thousands of square feet
  \item Constructed Area. constructed area in thousands of square feet
  \item Age. age of the property in years
}
}
\references{
Narula, S. C. and Wellington, J. F. (1977). Prediction, linear regression and minimum sum
of absolute errors. \emph{Technometrics}, \strong{19}, 185-190. \doi{10.2307/1268628}
}
\keyword{datasets}
