% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodD.R
\name{likelihoodD}
\alias{likelihoodD}
\title{Calculate Likelihood Displacement}
\usage{
likelihoodD(y, X)
}
\arguments{
\item{y}{A vector with response variables.}

\item{X}{A matrix or vector with explanatory variables.}
}
\value{
\item{Likelihood Displacement}{          A vector with Likelihood Displacement for each observation.}.
}
\description{
Calculate Likelihood Displacement
}
\examples{
### Using stackloss data

likelihoodD(stack.loss, stack.x)
}
\references{
Elian, S. N., André, C. D. S. and Narula, S. C. (2000) Influence Measure for the
\ifelse{html}{\out{L<sub>1</sub>}}{\eqn{L1}} regression.
\emph{Communications in Statistics - Theory and Methods}, \strong{29}(4), 837-849. \doi{10.1080/03610920008832518}.
}
